requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
requests==2.26.0
Pillow==8.4.0
PyPDF2==1.26.0